Remove cyclic dependencies, re-enable acyclic
